requirements.txt
crewai
crewai-tools
langchain-google-genai
python-dotenv
praw